+++ noodles-vcf/src/record/quality_score.rs
@@ -1,29 +1,18 @@
 //! VCF record quality score.
 
-use std::{error, fmt, num, ops::Deref, str::FromStr};
+use std::{error, fmt, num, str::FromStr};
 
-use super::{value::parse_f32_case_insensitive_extended, MISSING_FIELD};
+use super::value::parse_f32_case_insensitive_extended;
 
 const MIN: f32 = 0.0;
 
 /// A VCF record quality score.
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct QualityScore(Option<f32>);
-
-impl Deref for QualityScore {
-    type Target = Option<f32>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct QualityScore(f32);
 
 impl fmt::Display for QualityScore {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self.0 {
-            Some(score) => write!(f, "{}", score),
-            None => f.write_str(MISSING_FIELD),
-        }
+        write!(f, "{}", self.0)
     }
 }
 
@@ -51,7 +40,7 @@ impl TryFrom<f32> for QualityScore {
         if value < MIN {
             Err(TryFromFloatError::Negative)
         } else {
-            Ok(Self(Some(value)))
+            Ok(Self(value))
         }
     }
 }
@@ -85,11 +74,16 @@ impl FromStr for QualityScore {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "" => Err(ParseError::Empty),
-            MISSING_FIELD => Ok(Self(None)),
             _ => parse_f32_case_insensitive_extended(s)
                 .map_err(ParseError::Invalid)
                 .and_then(|value| Self::try_from(value).map_err(ParseError::InvalidValue)),
         }
+    }
+}
+
+impl From<QualityScore> for f32 {
+    fn from(quality_score: QualityScore) -> Self {
+        quality_score.0
     }
 }
 
@@ -99,17 +93,14 @@ mod tests {
 
     #[test]
     fn test_fmt() {
-        let quality_score = QualityScore(Some(1.3));
+        let quality_score = QualityScore(1.3);
         assert_eq!(quality_score.to_string(), "1.3");
-
-        let quality_score = QualityScore(None);
-        assert_eq!(quality_score.to_string(), ".");
     }
 
     #[test]
     fn test_try_from_float_for_quality_score() {
-        assert_eq!(QualityScore::try_from(0.0), Ok(QualityScore(Some(0.0))));
-        assert_eq!(QualityScore::try_from(13.0), Ok(QualityScore(Some(13.0))));
+        assert_eq!(QualityScore::try_from(0.0), Ok(QualityScore(0.0)));
+        assert_eq!(QualityScore::try_from(13.0), Ok(QualityScore(13.0)));
 
         assert_eq!(
             QualityScore::try_from(-8.0),
@@ -119,11 +110,14 @@ mod tests {
 
     #[test]
     fn test_from_str() {
-        assert_eq!(".".parse(), Ok(QualityScore(None)));
-        assert_eq!("5.8".parse(), Ok(QualityScore(Some(5.8))));
-        assert_eq!("Infinity".parse(), Ok(QualityScore(Some(f32::INFINITY))));
+        assert_eq!("5.8".parse(), Ok(QualityScore(5.8)));
+        assert_eq!("Infinity".parse(), Ok(QualityScore(f32::INFINITY)));
 
         assert_eq!("".parse::<QualityScore>(), Err(ParseError::Empty));
+        assert!(matches!(
+            ".".parse::<QualityScore>(),
+            Err(ParseError::Invalid(_))
+        ));
         assert!(matches!(
             "ndls".parse::<QualityScore>(),
             Err(ParseError::Invalid(_))
@@ -132,5 +126,10 @@ mod tests {
             "-8.5".parse::<QualityScore>(),
             Err(ParseError::InvalidValue(_))
         ));
+    }
+
+    #[test]
+    fn test_from_quality_score_for_f32() {
+        assert_eq!(f32::from(QualityScore(13.0)).to_bits(), (13.0f32).to_bits());
     }
 }